[move] TokenManager util -> datasources
