+++ app/src/main/java/com/ssu/better/presentation/ui/splash/SplashViewModel.kt
@@ -2,7 +2,7 @@ package com.ssu.better.presentation.ui.splash
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.ssu.better.data.util.TokenManager
+import com.ssu.better.data.datasources.TokenManager
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
